Ajustes passando regras para camada sevice Autor

+++ src/main/java/com/afzdev/demo/livraria/service/impl/AutorServiceImpl.java
@@ -1,15 +1,19 @@
 package com.afzdev.demo.livraria.service.impl;
 
 import com.afzdev.demo.livraria.dto.AutorDTO;
+import com.afzdev.demo.livraria.entities.Autor;
+import com.afzdev.demo.livraria.exceptions.DataBaseException;
+import com.afzdev.demo.livraria.exceptions.ResourceNotFoundException;
 import com.afzdev.demo.livraria.mapper.AutorMapper;
 import com.afzdev.demo.livraria.repository.AutorRepository;
 import com.afzdev.demo.livraria.service.AutorService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class AutorServiceImpl implements AutorService {
@@ -33,10 +37,12 @@ public class AutorServiceImpl implements AutorService {
 
     @Override
     @Transactional(readOnly = true)
-    public Optional<AutorDTO> buscarPorId(Long id) {
-        return Optional.of(autorMapper.toDTO(autorRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Não foi encontrado Autor com id : "+id))));
+    public AutorDTO buscarPorId(Long id) {
+        return autorRepository.findById(id)
+                .map(autorMapper::toDTO)
+                .orElseThrow(() -> new ResourceNotFoundException("Não foi encontrado Autor com id : "+id));
     }
+
 
     @Override
     @Transactional
@@ -45,15 +51,32 @@ public class AutorServiceImpl implements AutorService {
     }
 
     @Override
-    @Transactional
-    public void excluir(Long id) {
-        autorRepository.deleteById(id);
+    public AutorDTO atualizar(Long id, AutorDTO autorDto) {
+        try {
+            Autor autor = autorRepository.getReferenceById(id);
+            updateToEntity(autorDto,autor);
+            return autorMapper.toDTO(autorRepository.save(autor));
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Não foi encontrado Autor com id : "+id);
+        }
     }
 
-
-
     @Override
-    public boolean existePorId(Long id) {
-        return autorRepository.existsById(id);
+    @Transactional
+    public void excluir(Long id) {
+        try{
+            if (!autorRepository.existsById(id)){
+                throw new ResourceNotFoundException("Não foi encontrado Autor com id ao tentar excluir: "+id);
+            }
+            autorRepository.deleteById(id);
+        }catch (
+                DataIntegrityViolationException e){
+            throw new DataBaseException("Falha de integridade referencial");
+        }
+    }
+
+    //auxiliares
+    private void updateToEntity(AutorDTO dto, Autor entity){
+        entity.setNome(dto.getNome());
     }
 }
